Conflicts:
	src/controller/RequestMapping.java
	src/controller/user/LoginController.java
	src/controller/user/LogoutController.java
	src/controller/user/RegisterUserController.java
	src/model/dao/UserDAO.java
	src/model/service/UserManager.java

+++ src/controller/article/DeleteMusicController.java
@@ -2,6 +2,7 @@ package controller.article;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import controller.Controller;
 import model.dao.MusicDAO;
@@ -13,9 +14,15 @@ public class DeleteMusicController implements Controller {
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
 		// 작성자와 로그인 아이디 검사 추가
-		
+		HttpSession session = request.getSession();
+		String userId = (String) session.getAttribute("userId"); 
 		int musicId = Integer.parseInt(request.getParameter("musicId"));
+	
+		if (!musicDAO.isArticleWriter(musicId, userId))
+						return "/article/articleMain.jsp";
+		
 		musicDAO.deleteMusicArticle(musicId);
+		musicDAO.deleteMusic(musicId);
 		
 		return "redirect:/article/articleMain.jsp";
 	}
